shopTHAT_V1/graph_rag_backend/pyproject.toml
# pyproject.toml
[tool.poetry]
name = "graphdb"
version = "0.1.0"
description = "Standalone tool to ingest a JSON manifest and store a keyword hierarchy in Neo4j"
authors = ["shopTHAT"]
license = "MIT"

# point Poetry at your package in src/backend
packages = [
  { include = "backend", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
neo4j = "^5.10.0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.2"
sentence-transformers = "^4.1.0"
faiss-cpu = "^1.11.0"
fastapi = "^0.115.13"
uvicorn = "^0.34.3"
pandas = "^2.3.0"
openpyxl = "^3.1.5"
pydantic = "^2.11.7"
email-validator = "^2.2.0"
groq = "^0.31.0"

[tool.poetry.scripts]
graphdb = "backend.graphdb:main"
start-server = "run_server:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
